refactor(media): New files for select-media-panel
